+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 26 January 2023
+- Optimized Weapon tooltip Scan for Temporary Weapon Enchants
+
 # 14 January 2023
 - Fixed bug causing Slam's cooldown, when combined with the Storm of Swords talent, to not be testable for conditions.